Refactored common print view functionality into parent class
