+++ src/modules/print/SettingsPrintView.js
@@ -18,407 +18,44 @@
 /* eslint-disable lodash/prefer-lodash-method */
 
 import _ from 'lodash';
-import PdfTable from 'voilab-pdf-table';
-import PdfTableFitColumn from 'voilab-pdf-table/plugins/fitcolumn';
+
+import PrintView from './PrintView';
 
 import {
-  getTimezoneFromTimePrefs,
-  formatBirthdate,
-  formatCurrentDate,
-} from '../../utils/datetime';
+  getDeviceMeta,
+} from '../../utils/settings/data';
 
-import { getPatientFullName } from '../../utils/misc';
-
-const logo = require('./images/tidepool-logo-408x46.png');
-
-class SettingsPrintView {
+class SettingsPrintView extends PrintView {
   constructor(doc, data, opts) {
-    this.doc = doc;
+    super(doc, data, opts);
 
-    this.debug = opts.debug || false;
-
-    this.data = data;
-
-    this.dpi = opts.dpi;
-
-    this.margins = opts.margins;
-
-    this.font = 'Helvetica';
-    this.boldFont = 'Helvetica-Bold';
-
-    this.bgAxisFontSize = 5;
-    this.carbsFontSize = 5.5;
-    this.defaultFontSize = opts.defaultFontSize;
-    this.footerFontSize = opts.footerFontSize;
-    this.headerFontSize = opts.headerFontSize;
-    this.summaryHeaderFontSize = opts.summaryHeaderFontSize;
-
-    this.bgPrefs = opts.bgPrefs;
-    this.bgUnits = opts.bgPrefs.bgUnits;
-    this.bgBounds = opts.bgPrefs.bgBounds;
-    this.timePrefs = opts.timePrefs;
-    this.timezone = getTimezoneFromTimePrefs(opts.timePrefs);
-
-    this.width = opts.width;
-    this.height = opts.height;
-
-    this.patient = opts.patient;
-    this.patientInfoBox = {
-      width: 0,
-      height: 0,
-    };
-
-    this.colors = {
-      low: '#FF8B7C',
-      target: '#76D3A6',
-      basal: '#19A0D7',
-      high: '#BB9AE7',
-    };
-
-    this.rightEdge = this.margins.left + this.width;
-    this.bottomEdge = this.margins.top + this.height;
-
-    this.gapBtwnSummaryAndChartAsPercentage = 0.04;
-    this.chartArea = {
-      bottomEdge: opts.margins.top + opts.height,
-      leftEdge: opts.margins.left,
-      topEdge: opts.margins.top,
-    };
-
-    this.chartArea.width = this.rightEdge - this.chartArea.leftEdge;
-    this.initialChartArea = _.clone(this.chartArea);
-
-    this.totalPages = this.initialTotalPages = this.doc.bufferedPageRange().count || 0;
-
-    // kick off the dynamic calculation of chart area based on font sizes for header and footer
-    this.setHeaderSize().setFooterSize();
-
-    // no auto-binding :/
-    this.newPage = this.newPage.bind(this);
-    this.doc.on('pageAdded', this.newPage);
-  }
-
-  newPage() {
-    if (this.debug) {
-      this.renderDebugGrid();
-    }
-    this.renderHeader().renderFooter();
-  }
-
-  renderEventPath(path) {
-    if (path.type === 'programmed') {
-      this.doc.path(path.d)
-        .lineWidth(0.5)
-        .dash(0.5, { space: 1 })
-        .stroke(this.colors.bolus[path.type]);
-    } else {
-      this.doc.path(path.d)
-        .fill(this.colors.bolus[path.type]);
-    }
+    this.manufacturer = _.get(data, 'source', '').toLowerCase();
+    this.deviceMeta = getDeviceMeta(data, opts.timePrefs);
   }
 
   render() {
+    // console.log('data', this.data);
+    // console.log('deviceMeta', this.deviceMeta);
     this.doc.addPage();
-    this.renderOuterTable();
+    this.renderDeviceMeta();
+    this.renderBasalSettings();
   }
 
-  renderOuterTable() {
-    const table = new PdfTable(this.doc, {
-      bottomMargin: 30,
-    });
-
-    table.pdf.x = this.chartArea.leftEdge;
-    table.pdf.y = this.chartArea.topEdge;
-
-    table
-      // add some plugins (here, a 'fit-to-width' for a column)
-      .addPlugin(new PdfTableFitColumn({
-        column: 'description',
-      }))
-      // set defaults to your columns
-      .setColumnsDefaults({
-        headerBorder: 'B',
-        align: 'right',
-      })
-      // add table columns
-      .addColumns([
-        {
-          id: 'description',
-          header: 'Product',
-          align: 'left',
-          cache: false,
-        },
-        {
-          id: 'quantity',
-          header: 'Quantity',
-          width: 50,
-          cache: false,
-        },
-        {
-          id: 'price',
-          header: 'Price',
-          width: 40,
-          cache: false,
-        },
-        {
-          id: 'total',
-          header: 'Total',
-          width: 90,
-          height: this.renderInnerTable(),
-          cache: false,
-          renderer: (tb, data, draw, column, pos) => {
-            if (draw) {
-              this.renderInnerTable(data, pos);
-            }
-            return '';
-          },
-        },
-      ]);
-
-    table.addBody([
-      { description: 'Product 1', quantity: 1, price: 20.10, total: 20.10 },
-      { description: 'Product 2', quantity: 4, price: 4.00, total: 16.00 },
-      { description: 'Product 3', quantity: 2, price: 17.85, total: 35.70 },
-    ]);
-  }
-
-  renderInnerTable(data = {}, pos) {
-    const table = new PdfTable(this.doc, {
-      bottomMargin: 30,
-    });
-
-    if (pos) {
-      this.doc.x = pos.x;
-      this.doc.y = pos.y;
-    }
-
-    const initY = this.doc.y;
-
-    this.doc.text('whazzup?', this.doc.x, this.doc.y, {
-      align: 'right',
-    });
-    this.doc.moveDown(2);
-
-    table
-      .setShowHeaders(false)
-      .addPlugin(new PdfTableFitColumn({
-        column: 'price',
-      }))
-      // set defaults to your columns
-      .setColumnsDefaults({
-        align: 'right',
-      })
-      // add table columns
-      .addColumns([
-        {
-          id: 'spacer',
-          header: '',
-          width: this.doc.x - this.chartArea.leftEdge,
-        },
-        {
-          id: 'price',
-          header: 'Price',
-          // width: 30,
-        },
-        {
-          id: 'total',
-          header: 'Total',
-          width: 30,
-          renderer: (tb2, innerData) => `CHF ${innerData.total || ''}`,
-        },
-      ]);
-
-    table.addBody([
-      { price: data.price, total: data.total },
-    ]);
-
-    return this.doc.y - initY;
-  }
-
-  renderPatientInfo() {
-    const patientName = getPatientFullName(this.patient);
-    const patientBirthdate = formatBirthdate(this.patient);
-    const xOffset = this.margins.left;
-    const yOffset = this.margins.top;
-
+  renderDeviceMeta() {
     this.doc
-      .lineWidth(1)
-      .fontSize(10)
-      .text(patientName, xOffset, yOffset, {
-        lineGap: 2,
-      });
+      .font(this.boldFont)
+      .fontSize(this.defaultFontSize)
+      .text(this.data.source, { continued: true })
+      .font(this.font)
+      .text(` Uploaded on ${this.deviceMeta.uploaded}`, { continued: true })
+      .text(` › Serial Number: ${this.deviceMeta.serial}`)
+      .moveDown();
 
-    const patientNameWidth = this.patientInfoBox.width = this.doc.widthOfString(patientName);
-
-    this.doc
-      .fontSize(10)
-      .text(patientBirthdate);
-
-    const patientBirthdayWidth = this.doc.widthOfString(patientBirthdate);
-    this.patientInfoBox.height = this.doc.y;
-
-    if (patientNameWidth < patientBirthdayWidth) {
-      this.patientInfoBox.width = patientBirthdayWidth;
-    }
-
-    // Render the divider between the patient info and title
-    const padding = 10;
-
-    this.doc
-      .moveTo(this.margins.left + this.patientInfoBox.width + padding, this.margins.top)
-      .lineTo(this.margins.left + this.patientInfoBox.width + padding, this.patientInfoBox.height)
-      .stroke('black');
-
-    this.dividerWidth = padding * 2 + 1;
+    this.resetText();
   }
 
-  renderTitle() {
-    const title = 'Pump Settings';
-    const lineHeight = this.doc.fontSize(14).currentLineHeight();
-    const xOffset = this.margins.left + this.patientInfoBox.width + 21;
-    const yOffset = (
-      this.margins.top + ((this.patientInfoBox.height - this.margins.top) / 2 - (lineHeight / 2))
-    );
-
-    this.doc.text(title, xOffset, yOffset);
-    this.titleWidth = this.doc.widthOfString(title);
-  }
-
-  renderPrintDate() {
-    const lineHeight = this.doc.fontSize(14).currentLineHeight();
-
-    // Calculate the remaining available width so we can
-    // center the print text between the patient/title text and the logo
-    const availableWidth = this.doc.page.width - _.reduce([
-      this.patientInfoBox.width,
-      this.dividerWidth,
-      this.titleWidth,
-      this.logoWidth,
-      this.margins.left,
-      this.margins.right,
-    ], (a, b) => (a + b), 0);
-
-    const xOffset = (
-      this.margins.left + this.patientInfoBox.width + this.dividerWidth + this.titleWidth
-    );
-    const yOffset = (
-      this.margins.top + ((this.patientInfoBox.height - this.margins.top) / 2 - (lineHeight / 2))
-    );
-
-    this.doc
-      .fontSize(10)
-      .text(`Printed from Tidepool: ${formatCurrentDate()}`, xOffset, yOffset + 4, {
-        width: availableWidth,
-        align: 'center',
-      });
-  }
-
-  renderLogo() {
-    this.logoWidth = 100;
-    const xOffset = this.doc.page.width - this.logoWidth - this.margins.right;
-    const yOffset = this.margins.top + 6;
-
-    this.doc.image(logo, xOffset, yOffset, { width: this.logoWidth });
-  }
-
-  renderDebugGrid() {
-    const minorLineColor = '#B8B8B8';
-    const numMinorLines = 5;
-    let thisLineYPos = this.margins.top;
-    while (thisLineYPos <= (this.bottomEdge)) {
-      this.doc.moveTo(this.margins.left, thisLineYPos)
-        .lineTo(this.rightEdge, thisLineYPos)
-        .lineWidth(0.25)
-        .stroke('red');
-      if (thisLineYPos !== this.bottomEdge) {
-        for (let i = 1; i < numMinorLines + 1; ++i) {
-          const innerLinePos = thisLineYPos + this.dpi * (i / (numMinorLines + 1));
-          this.doc.moveTo(this.margins.left, innerLinePos)
-            .lineTo(this.rightEdge, innerLinePos)
-            .lineWidth(0.05)
-            .stroke(minorLineColor);
-        }
-      }
-      thisLineYPos += this.dpi;
-    }
-
-    let thisLineXPos = this.margins.left;
-    while (thisLineXPos <= (this.rightEdge)) {
-      this.doc.moveTo(thisLineXPos, this.margins.top)
-        .lineTo(thisLineXPos, this.bottomEdge)
-        .lineWidth(0.25)
-        .stroke('red');
-      for (let i = 1; i < numMinorLines + 1; ++i) {
-        const innerLinePos = thisLineXPos + this.dpi * (i / (numMinorLines + 1));
-        if (innerLinePos <= this.rightEdge) {
-          this.doc.moveTo(innerLinePos, this.margins.top)
-            .lineTo(innerLinePos, this.bottomEdge)
-            .lineWidth(0.05)
-            .stroke(minorLineColor);
-        }
-      }
-      thisLineXPos += this.dpi;
-    }
-
-    return this;
-  }
-
-  renderHeader() {
-    this.renderPatientInfo();
-
-    this.renderTitle();
-
-    this.renderLogo();
-
-    this.renderPrintDate();
-
-    this.doc.moveDown();
-
-    const lineHeight = this.doc.fontSize(14).currentLineHeight();
-    const height = lineHeight * 2.25 + this.margins.top;
-    this.doc
-      .moveTo(this.margins.left, height)
-      .lineTo(this.margins.left + this.width, height)
-      .stroke('black');
-
-    // TODO: remove this; it is just for exposing/debugging the chartArea.topEdge adjustment
-    if (this.debug) {
-      this.doc
-        .fillColor('#E8E8E8', 0.3333333333)
-        .rect(this.margins.left, this.margins.top, this.width, lineHeight * 4)
-        .fill();
-    }
-
-    return this;
-  }
-
-  renderFooter() {
-    const lineHeight = this.doc.fontSize(10).currentLineHeight();
-    const helpText = 'Questions or feedback please email support@tidepool.org ' +
-                     'or visit support.tidepool.org';
-
-    this.doc.fillColor('black').fillOpacity(1)
-      .text(helpText, this.margins.left, this.bottomEdge - lineHeight * 1.4, {
-        align: 'center',
-      });
-
-    return this;
-  }
-
-  setFooterSize() {
-    this.doc.fontSize(this.footerFontSize);
-    const lineHeight = this.doc.currentLineHeight();
-    this.chartArea.bottomEdge = this.chartArea.bottomEdge - lineHeight * 9;
-
-    return this;
-  }
-
-  setHeaderSize() {
-    this.doc.fontSize(this.headerFontSize);
-    const lineHeight = this.doc.currentLineHeight();
-    this.chartArea.topEdge = this.chartArea.topEdge + lineHeight * 4;
-
-    return this;
+  renderBasalSettings() {
+    this.renderSectionHeading('Basal Rates');
   }
 }
